refactor: use factory pattern on application and projection layers
